requirements.txt
openai
anthropic
pandas
tenacity
mangoengine
click-8.1.7
joblib-1.4.2
nltk-3.9.1
regex-2024.11.6
scikit-learn-1.5.2
scipy-1.13.1
threadpoolctl-3.5.0
contourpy-1.3.0
cycler-0.12.1
fonttools-4.55.0
importlib-resources-6.4.5
kiwisolver-1.4.7
matplotlib-3.9.2
pillow-11.0.0
pyparsing-3.2.0
matplotlib-venn-1.1.1